refactor: separate dataBase
